docker/tutorials/2-compose/requirements.txt
click==8.1.8
Flask==2.0.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
mysql-connector-python==8.0.26
protobuf==6.30.2
Werkzeug==2.0.3
